+++ src/pbrt/euclidean_space/vector2.h
@@ -10,7 +10,7 @@ class Vector2 {
 
     PBRT_CPU_GPU Vector2(T _x, T _y) : x(_x), y(_y) {}
 
-    PBRT_CPU_GPU T &operator[](int index) {
+    PBRT_CPU_GPU T &operator[](uint8_t index) {
         switch (index) {
         case 0: {
             return x;
